+++ features/support/app_pages/marketplace/mp_group_affiliations_page.rb
@@ -28,4 +28,8 @@ class MarketplaceGroupAffiliationsPage < IDmeBase
 	def click_verify_government_affiliation_link
 		find(:link, :text =>"Government Employees").click
 	end 
+
+	def click_verify_student_affiliation_link
+		find(:link, :text =>"Students").click
+	end 
 end